requirements.txt
python-fasthtml>=0.1.0
django>=4.2.0
django-machina>=1.2.0  # Forum framework
django-haystack>=3.2.1  # Required by django-machina for search
django-widget-tweaks>=1.5.0  # Required by django-machina for form rendering
Whoosh>=2.7.4  # Search backend for django-haystack
Pillow>=10.0.0  # Required for image handling
python-dotenv>=1.0.0
httpx>=0.24.0
starlette>=0.27.0
uvicorn>=0.23.0
python-multipart>=0.0.6
itsdangerous>=2.1.2  # for session handling
django-allauth>=0.57.0  # for OAuth (GitHub, GitLab)
